Add elevation profile in NetCDF output
[skip ci][ci skip]
