+++ src/diagnostics/Diagnostics.jl
@@ -5,7 +5,7 @@ import ClimaComms
 import LinearAlgebra: dot
 
 import ClimaCore:
-    Fields, Geometry, InputOutput, Meshes, Spaces, Operators, Domains
+    Fields, Geometry, InputOutput, Meshes, Spaces, Operators, Domains, Grids
 import ClimaCore.Utilities: half
 import Thermodynamics as TD
